Make this project a library project
The only way I can think of now to decouple playing logic with the
rule system itself is to make this officially a library project.
The engine will be thus providing game status transition but nothing
really media-related.

That being said, I'm considering adding `cursor` as a basic thing
in the descriptor...