Fix initial history condition
It doesn't make sense to prepend the "FF[4]..." info for all loaded
SGF. Removing it and also apply to_sgf_node to shift g.history
for the initial loading. Previously I was intended to use to_tail
and append methods for this purpose, but it turns out very tedious
and the TreeNode struct has corner cases that is too dirty to
handle.