Refactor is_illegal_setup1
The bool return value is no longer suitable, so changing it into
more action-like form, returning Result.

Notably, test_setup1_4 should panic with the newly added error, because
we now change the testing frequency from **when setup1 is considered
done** to **each time an attempt to add a marker**. The previous
error is no longer possible due to Pigeonhole Principle.