[core::action] Prepare for the later refactor to add_single_marker
Previously the `add_single_marker` method was not as stand-alone as
other `add_` functions. In fact, it is the most complicated one.
Before this refactor, a given coordinate is just added without
difficult checks. The thorough check is only performed after all
markers are put.

After we have this `marker_slot` with the capacity information,
we can truely transform `add-single-marker` afterward.