[core] Fix the bug in viable
It was a short-cut pattern that if any one between them works then
returns true, and return false at the end because it never finds
a way out. However, the previous commit adds two nonviable patterns
and didn't obey to the principle above, so many problems occured.

Also, the iter() to HashMap does not guarentee order. Note this.