Fix the ending condition
I didn't handle the ending condition well. There was `end()` to
immutably check the game status, and only checks the "current
player". This can go wrong like the previous coord_server, in which
`next()` is always called before `end`, and thus causing the delay
of the ending announcement.

With this fix, the new phase, `Phase::End` will be checked in an
immutable function `is_ended`, and it is set in `check_and_set_end`
function, triggered by the end of any `commit_action` calls except
for the skipped ones.