Fix the setup state machine
Two folds:
1. The position check of steup2 was too optimistic. It was possible
that the random-generated env position overwirtes the character of
the same world, causing later trouble that is not possible to be
cleaned up. We fix this problem by checking the attempting env
position in both load_history and setup_with_coord. This problem was
discovered with seed d94ce507c72b47f4960187d0f2dfab9c.
2. The turn switch states are not consistent between the mentioned
two functions. We sync them up and update a few test cases for them
as well.