action: fix doctor status check
There were previously two bugs.

1. Sorting each marker_slot in the wrong order.
2. When there are more encountered plague markers than 5, no strict
   order is required by the rule.

However, directly applying `doctor_is_free` to the current Coord
doesn't work because the function was only used for updating
`marker_slot` so far, and it can not express the semantics for the
candidate coord. As a result, generalize the need to reuse the
conditions in that function by adding one boolean variable.